+++ Sources/MapboxMaps/Gestures/GestureManager.swift
@@ -28,7 +28,6 @@ public final class GestureManager: GestureHandlerDelegate {
             quickZoomGestureRecognizer.isEnabled = newValue.quickZoomEnabled
             panGestureHandler.panMode = newValue.panMode
             panGestureHandler.decelerationFactor = newValue.panDecelerationFactor
-            animationLockoutGestureHandler.gestureRecognizer.isEnabled = newValue.disableAnimationsDuringGestures
         }
         get {
             var gestureOptions = GestureOptions()
@@ -42,7 +41,6 @@ public final class GestureManager: GestureHandlerDelegate {
             gestureOptions.quickZoomEnabled = quickZoomGestureRecognizer.isEnabled
             gestureOptions.panMode = panGestureHandler.panMode
             gestureOptions.panDecelerationFactor = panGestureHandler.decelerationFactor
-            gestureOptions.disableAnimationsDuringGestures = animationLockoutGestureHandler.gestureRecognizer.isEnabled
             return gestureOptions
         }
     }
@@ -85,6 +83,10 @@ public final class GestureManager: GestureHandlerDelegate {
         return singleTapGestureHandler.gestureRecognizer
     }
 
+    internal var anyTouchGestureRecognizer: UIGestureRecognizer {
+        return anyTouchGestureHandler.gestureRecognizer
+    }
+
     /// Set this delegate to be called back if a gesture begins
     public weak var delegate: GestureManagerDelegate?
 
@@ -95,7 +97,7 @@ public final class GestureManager: GestureHandlerDelegate {
     private let doubleTouchToZoomOutGestureHandler: GestureHandler
     private let quickZoomGestureHandler: GestureHandler
     private let singleTapGestureHandler: GestureHandler
-    private let animationLockoutGestureHandler: GestureHandler
+    private let anyTouchGestureHandler: GestureHandler
     private let mapboxMap: MapboxMapProtocol
 
     internal init(panGestureHandler: PanGestureHandlerProtocol,
@@ -105,7 +107,7 @@ public final class GestureManager: GestureHandlerDelegate {
                   doubleTouchToZoomOutGestureHandler: GestureHandler,
                   quickZoomGestureHandler: GestureHandler,
                   singleTapGestureHandler: GestureHandler,
-                  animationLockoutGestureHandler: GestureHandler,
+                  anyTouchGestureHandler: GestureHandler,
                   mapboxMap: MapboxMapProtocol) {
         self.panGestureHandler = panGestureHandler
         self.pinchGestureHandler = pinchGestureHandler
@@ -114,7 +116,7 @@ public final class GestureManager: GestureHandlerDelegate {
         self.doubleTouchToZoomOutGestureHandler = doubleTouchToZoomOutGestureHandler
         self.quickZoomGestureHandler = quickZoomGestureHandler
         self.singleTapGestureHandler = singleTapGestureHandler
-        self.animationLockoutGestureHandler = animationLockoutGestureHandler
+        self.anyTouchGestureHandler = anyTouchGestureHandler
         self.mapboxMap = mapboxMap
 
         panGestureHandler.delegate = self