+++ Tests/MapboxMapsTests/Gestures/GestureOptionsTests.swift
@@ -15,6 +15,5 @@ final class GestureOptionsTests: XCTestCase {
         XCTAssertTrue(options.quickZoomEnabled)
         XCTAssertEqual(options.panMode, .horizontalAndVertical)
         XCTAssertEqual(options.panDecelerationFactor, UIScrollView.DecelerationRate.normal.rawValue)
-        XCTAssertTrue(options.disableAnimationsDuringGestures)
     }
 }